webapp/pyproject.toml
[project]
name = "ref-webapp"
version = "0.1.0"
description = "Web application dependencies for REF"
requires-python = ">=3.10"
dependencies = [
    "ansi2html==1.9.2",
    "argh==0.31.3",
    "arrow==1.3.0",
    "async-timeout==5.0.1",
    "backports.tarfile==1.2.0",
    "cffi==1.17.1",
    "coloredlogs==15.0.1",
    "docker==7.1.0",
    "flask-bcrypt==1.0.1",
    "flask-debugtoolbar==0.16.0",
    "flask-failsafe==0.2",
    "flask-limiter==3.10.1",
    "flask-login==0.6.3",
    "flask-migrate==4.1.0",
    "flask-moment==1.0.6",
    "fuzzywuzzy==0.18.0",
    "PySocks @ git+https://github.com/nbars/PySocks.git@hack_unix_domain_socket_file_support",
    "gunicorn==23.0.0",
    "hypothesis==6.124.7",
    "importlib-metadata==8.6.1",
    "jaraco.collections==5.1.0",
    "pip-chill==1.0.3",
    "platformdirs==4.2.2",
    "psycopg2-binary==2.9.10",
    "py==1.11.0",
    "pycryptodome==3.21.0",
    "pyparsing==3.2.1",
    "python-levenshtein==0.26.1",
    "python-telegram-handler==2.2.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "rq==2.1.0",
    "toml==0.10.2",
    "tomli==2.2.1",
    "uwsgi==2.0.28",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "wtforms==3.2.1",
]
